tools: Move task logic to buildSrc
Move the task logic to the buildSrc project for better organization.
