+++ tools/curations/buildSrc/src/main/kotlin/Utils.kt
@@ -1,0 +1,62 @@
+package org.ossreviewtoolkit.tools.curations
+
+import com.vdurmont.semver4j.Semver
+import org.gradle.api.DefaultTask
+import org.gradle.api.logging.Logging
+import org.ossreviewtoolkit.model.Identifier
+import org.ossreviewtoolkit.model.PackageCuration
+import org.ossreviewtoolkit.model.PackageCurationData
+import org.ossreviewtoolkit.model.VcsInfoCurationData
+import org.ossreviewtoolkit.utils.common.encodeOr
+import java.io.File
+
+private val logger = Logging.getLogger("utils")
+private val versionRegex = Regex("^v?\\d+\\.\\d+\\.\\d+\$")
+
+val DefaultTask.curationsDir: File
+    get() = project.rootDir.resolve("../../curations")
+
+fun Identifier.toCurationPath() =
+    "${type.encodeOr("_")}/${namespace.encodeOr("_")}/${name.encodeOr("_")}.yml"
+
+fun List<PathCurationData>.toPathCurations(): List<PackageCuration> {
+    var lastPath = ""
+
+    // Group subsequent versions which belong to the same path.
+    val grouped = sortedBy { Semver(it.tag.removePrefix("v")) }
+        .fold(mutableListOf<MutableList<PathCurationData>>()) { acc, cur ->
+            if (cur.path != lastPath) {
+                acc.add(mutableListOf())
+            }
+            lastPath = cur.path
+            acc.apply { last() += cur }
+        }
+
+    return grouped.mapIndexed { index, curations ->
+        val firstVersion = curations.first().tag.removePrefix("v")
+        val lastVersion = curations.last().tag.removePrefix("v")
+
+        val versionRange = when {
+            index == grouped.size - 1 -> "[$firstVersion,)"
+            firstVersion == lastVersion -> firstVersion
+            else -> "[$firstVersion,$lastVersion]"
+        }
+
+        val id = curations.first().id.copy(version = versionRange)
+        val path = curations.first().path
+
+        logger.quiet("Creating path curation for id=${id.toCoordinates()} path=$path.")
+
+        PackageCuration(
+            id = id.copy(version = versionRange),
+            data = PackageCurationData(
+                comment = "Set the VCS path of the module inside the multi-module repository.",
+                vcs = VcsInfoCurationData(
+                    path = path
+                )
+            )
+        )
+    }
+}
+
+fun String.isVersion() = matches(versionRegex)
